♻️🧪 (Refactoring and testing) - Move all files to a lib folder + Package the extension to test if everything is working properly before publishing it
