+++ .vscodeignore
@@ -9,3 +9,7 @@ vsc-extension-quickstart.md
 **/*.map
 **/*.ts
 **/.vscode-test.*
+
+# Custom rules
+!src/lib/assets/
+!src/lib/components/
