+++ src/lib/assets/images/_icon.png
