+++ src/lib/assets/images/captures/sidebar.png
