+++ src/lib/assets/uri/styleCodiconUri.ts
