+++ src/lib/assets/uri/styleResetUri.ts
@@ -2,7 +2,14 @@ import * as vscode from "vscode";
 
 const styleResetUri = (webview: vscode.Webview, extensionUri: vscode.Uri) =>
   webview.asWebviewUri(
-    vscode.Uri.joinPath(extensionUri, "src", "assets", "styles", "reset.css")
+    vscode.Uri.joinPath(
+      extensionUri,
+      "src",
+      "lib",
+      "assets",
+      "styles",
+      "reset.css"
+    )
   );
 
 export default styleResetUri;
