+++ src/lib/assets/uri/styleTailwindUri.ts
@@ -1,13 +1,11 @@
 import * as vscode from "vscode";
 
-const styleTailwindUri = (
-  webview: vscode.Webview,
-  extensionUri: vscode.Uri
-) =>
+const styleTailwindUri = (webview: vscode.Webview, extensionUri: vscode.Uri) =>
   webview.asWebviewUri(
     vscode.Uri.joinPath(
       extensionUri,
       "src",
+      "lib",
       "assets",
       "styles",
       "tailwind.min.css"
