+++ src/lib/components/Documentation/uri/styleDocumentationUri.ts
@@ -1,0 +1,19 @@
+import * as vscode from "vscode";
+
+const styleDocumentationUri = (
+  webview: vscode.Webview,
+  extensionUri: vscode.Uri
+) =>
+  webview.asWebviewUri(
+    vscode.Uri.joinPath(
+      extensionUri,
+      "src",
+      "lib",
+      "components",
+      "Documentation",
+      "styles",
+      "documentation.css"
+    )
+  );
+
+export default styleDocumentationUri;
