+++ src/lib/components/Sidebar/scripts/sidebar.js
