+++ src/lib/interfaces/IPackageInformation.ts
