+++ src/lib/utils/getPackageInfo.ts
